@@ -1,34 +1,30 @@
 package zlk.ast;
 
-import java.util.function.Consumer;
-import java.util.function.Function;
-
+import zlk.ast.AType.Arrow;
+import zlk.ast.AType.Atom;
+import zlk.util.Location;
 import zlk.util.LocationHolder;
 import zlk.util.pp.PrettyPrintable;
+import zlk.util.pp.PrettyPrinter;
 
 public sealed interface AType extends PrettyPrintable, LocationHolder
-permits ATyAtom, ATyArrow {
-	default <R> R fold(
-			Function<? super ATyAtom, ? extends R> forBase,
-			Function<? super ATyArrow, ? extends R> forArrow) {
-		if(this instanceof ATyAtom base) {
-			return forBase.apply(base);
-		} else if(this instanceof ATyArrow arrow) {
-			return forArrow.apply(arrow);
-		} else {
-			throw new Error(this.getClass().toString());
-		}
-	}
+permits Atom, Arrow {
+	record Atom(String name, Location loc) implements AType {};
+	record Arrow(AType arg, AType ret, Location loc) implements AType {};
 
-	default void match(
-			Consumer<? super ATyAtom> forBase,
-			Consumer<? super ATyArrow> forArrow) {
-		if(this instanceof ATyAtom base) {
-			forBase.accept(base);
-		} else if(this instanceof ATyArrow arrow) {
-			forArrow.accept(arrow);
-		} else {
-			throw new Error(this.getClass().toString());
+	@Override
+	default void mkString(PrettyPrinter pp) {
+		switch(this) {
+		case Atom atom -> {
+			pp.append(atom.name);
+		}
+		case Arrow arrow -> {
+			switch(arrow.arg) {
+			case Atom _ -> pp.append(arrow.arg);
+			case Arrow _ -> pp.append("(").append(arrow.arg).append(")");
+			}
+			pp.append(" -> ").append(arrow.ret);
+		}
 		}
 	}
 }
