@@ -1,36 +1,31 @@
 package zlk.ast;
 
-import java.util.function.Consumer;
-import java.util.function.Function;
+import java.util.List;
 
+import zlk.ast.Pattern.Ctor;
+import zlk.ast.Pattern.Var;
+import zlk.util.Location;
 import zlk.util.LocationHolder;
 import zlk.util.pp.PrettyPrintable;
+import zlk.util.pp.PrettyPrinter;
 
 public sealed interface Pattern extends PrettyPrintable, LocationHolder
-permits PVar, PCtor {
-	default <R> R fold(
-			Function<? super PVar, ? extends R> forVar,
-			Function<? super PCtor, ? extends R> forCtor) {
-		if(this instanceof PVar var) {
-			return forVar.apply(var);
-		} else if(this instanceof PCtor ctor) {
-			return forCtor.apply(ctor);
-		} else {
-			throw new Error(this.getClass().toString());
-		}
-	}
+permits Var, Ctor {
+	record Var(String name, Location loc) implements Pattern {}
+	record Ctor(String name, List<Pattern> args, Location loc) implements Pattern {}
 
-	default void match(
-			Consumer<? super PVar> forVar,
-			Consumer<? super PCtor> forCtor) {
-		if(this instanceof PVar var) {
-			forVar.accept(var);
-			return;
-		} else if(this instanceof PCtor ctor) {
-			forCtor.accept(ctor);
-			return;
-		} else {
-			throw new Error(this.getClass().toString());
+	@Override
+	default void mkString(PrettyPrinter pp) {
+		switch(this) {
+		case Var(String name, _) -> {
+			pp.append(name);
+		}
+		case Ctor(String name, List<Pattern> args, _) -> {
+			pp.append(name);
+			for(Pattern arg : args) {
+				pp.append(" ").append(arg);
+			}
+		}
 		}
 	}
 }
