@@ -1,6 +1,6 @@
 package zlk.idcalc;
 
-import zlk.common.cnst.ConstValue;
+import zlk.common.ConstValue;
 import zlk.common.type.TyAtom;
 import zlk.util.Location;
 import zlk.util.pp.PrettyPrinter;
@@ -11,9 +11,10 @@ public record IcCnst(
 implements IcExp {
 
 	public TyAtom type() {
-		return value().fold(
-				bool -> TyAtom.BOOL,
-				i32  -> TyAtom.I32);
+		return switch(value) {
+		case ConstValue.Bool _ -> TyAtom.BOOL;
+		case ConstValue.I32 _ -> TyAtom.I32;
+		};
 	}
 
 	@Override
